DATAES-488 - Extract ReactiveClient interface and add integration tests.
